refactor: validateRaceNumber 함수 모듈화
